+++ activejob/CHANGELOG.md
@@ -24,4 +24,12 @@
 
     *DHH*
 
+*   Added instance variable `@queue` to JobWrapper.
+
+    This will fix issues in [resque-scheduler](https://github.com/resque/resque-scheduler) `#job_to_hash` method,
+    so we can use `#enqueue_delayed_selection`, `#remove_delayed` method in resque-scheduler smoothly.
+
+    *mu29*
+
+
 Please check [5-0-stable](https://github.com/rails/rails/blob/5-0-stable/activejob/CHANGELOG.md) for previous changes.